+++ core/vnl/algo/vnl_adaptsimpson_integral.cxx
@@ -1,17 +1,16 @@
 #include "vnl_adaptsimpson_integral.h"
 #include <vnl/algo/vnl_netlib.h>
 
-double vnl_adaptsimpson_integral::int_fnct_(double* x)
+double vnl_adaptsimpson_integral::int_fnct_(float* x)
 {
   return  pfnct_->f_(*x);
 }
 
-double vnl_adaptsimpson_integral::integral(vnl_integrant_fnct* f, double a, 
-    double b, double acury)
+double vnl_adaptsimpson_integral::integral(vnl_integrant_fnct* f, float a, float b, float acury)
 {
 
-  double res = 0;
-  double srmat[1111], errbnd;
+  float res = 0;
+  float srmat[1111], errbnd;
 
   int m;
   int stat;
@@ -19,7 +18,7 @@ double vnl_adaptsimpson_integral::integral(vnl_integrant_fnct* f, double a,
   //set the function
   pfnct_ = f;
 
-  double tol = acury;
+  float tol = acury;
   
   adaptquad_(&vnl_adaptsimpson_integral::int_fnct_, &a, &b, &tol, srmat, &res, &errbnd, &m, &stat);
