+++ core/vnl/algo/vnl_simpson_integral.cxx
@@ -1,15 +1,15 @@
 #include "vnl_simpson_integral.h"
 #include <vnl/algo/vnl_netlib.h>
 
-double vnl_simpson_integral::int_fnct_(double* x)
+double vnl_simpson_integral::int_fnct_(float* x)
 {
   return  pfnct_->f_(*x);
 }
 
-double vnl_simpson_integral::integral(vnl_integrant_fnct* f, double a, double b, int n)
+double vnl_simpson_integral::integral(vnl_integrant_fnct* f, float a, float b, int n)
 {
 
-  double res = 0;
+  float res = 0;
  
   //set the function
   pfnct_ = f;
